probably fixed

+++ PymentGateway.PublishedLanguage/Events/ServiceCreated.cs
@@ -10,11 +10,11 @@ namespace PaymentGateway.PublishedLanguage.Events
     public class ServiceCreated: INotification
     {
         public string Name { get; set; }
-        public double Value { get; set; }
+        public decimal Value { get; set; }
         public int Limit { get; set; }
         public string Currency { get; set; }
 
-        public ServiceCreated(string name, double value, int limit, string currency)
+        public ServiceCreated(string name, decimal value, int limit, string currency)
         {
             this.Currency = currency;
             this.Limit = limit;
